# proxy/public/swagger.json
{
  "schemes": [
    "http"
  ],
  "swagger": "2.0",
  "info": {
    "description": "Documentation of Proxy API.",
    "title": "Proxy",
    "version": "1.0.0"
  },
  "host": "localhost:8080",
  "paths": {
    "/api/address/geocode": {
      "post": {
        "description": "Get full address info by coordinates",
        "produces": [
          "application/json"
        ],
        "operationId": "GeoCode",
        "parameters": [
          {
            "type": "string",
            "description": "latitude",
            "name": "lat",
            "in": "query",
            "required": true
          },
          {
            "type": "string",
            "description": "longitude",
            "name": "lng",
            "in": "query",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "full addresses",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/ResponseAddress"
              }
            }
          }
        }
      }
    },
    "/api/address/search": {
      "post": {
        "description": "Get full address info by its part",
        "produces": [
          "application/json"
        ],
        "operationId": "AddressSearch",
        "parameters": [
          {
            "type": "string",
            "description": "part of address",
            "name": "query",
            "in": "query",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "full addresses",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/ResponseAddress"
              }
            }
          }
        }
      }
    }
  },
  "definitions": {
    "Address": {
      "type": "object",
      "properties": {
        "city": {
          "type": "string",
          "x-go-name": "City"
        },
        "house": {
          "type": "string",
          "x-go-name": "House"
        },
        "lat": {
          "type": "string",
          "x-go-name": "Lat"
        },
        "lon": {
          "type": "string",
          "x-go-name": "Lon"
        },
        "street": {
          "type": "string",
          "x-go-name": "Street"
        }
      },
      "x-go-package": "test"
    },
    "ResponseAddress": {
      "type": "object",
      "required": [
        "addresses"
      ],
      "properties": {
        "addresses": {
          "description": "Addresses",
          "type": "array",
          "items": {
            "$ref": "#/definitions/Address"
          },
          "x-go-name": "Addresses"
        }
      },
      "x-go-package": "test"
    }
  }
}